requirements.txt
asgiref==3.11.1
Django==6.0.4
django-filter==25.2
djangorestframework==3.17.1
Markdown==3.10.2
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
typing_extensions==4.15.0
